Renaming VectorizedRDD to VeRDD.
